+++ packages/core/core/src/requests/BundleGraphRequest.js
@@ -72,10 +72,6 @@ type BundleGraphRequestInput = {|
   optionsRef: SharedReference,
 |};
 
-type BundleGraphRequestResult = {|
-  bundleGraph: InternalBundleGraph,
-|};
-
 type RunInput = {|
   input: BundleGraphRequestInput,
   ...StaticRunOpts<BundleGraphResult>,
@@ -84,6 +80,7 @@ type RunInput = {|
 // TODO: Rename to BundleGraphRequestResult
 export type BundleGraphResult = {|
   bundleGraph: InternalBundleGraph,
+  assetGraphBundlingVersion: number,
   changedAssets: Map<string, Asset>,
   assetRequests: Array<AssetGroup>,
 |};
@@ -215,6 +212,7 @@ export default function createBundleGraphRequest(
 
       if (subRequestsInvalid) {
         assetGraph.safeToIncrementallyBundle = false;
+        assetGraph.setNeedsBundling();
       }
 
       let configResult = nullthrows(
@@ -366,16 +364,17 @@ class BundlerRunner {
     let {plugin: bundler, name, resolveFrom} = plugin;
 
     // if a previous asset graph hash is passed in, check if the bundle graph is also available
-    let previousBundleGraphResult: ?BundleGraphRequestResult;
-    if (graph.safeToIncrementallyBundle) {
-      try {
-        previousBundleGraphResult = await this.api.getPreviousResult();
-      } catch {
-        // if the bundle graph had an error or was removed, don't fail the build
-      }
-    }
-    if (previousBundleGraphResult == null) {
+    const previousBundleGraphResult: ?BundleGraphResult =
+      await this.api.getPreviousResult();
+    const canIncrementallyBundle =
+      previousBundleGraphResult?.assetGraphBundlingVersion != null &&
+      graph.canIncrementallyBundle(
+        previousBundleGraphResult.assetGraphBundlingVersion,
+      );
+
+    if (graph.safeToIncrementallyBundle && previousBundleGraphResult == null) {
       graph.safeToIncrementallyBundle = false;
+      graph.setNeedsBundling();
     }
 
     let internalBundleGraph;
@@ -386,7 +385,7 @@ class BundlerRunner {
       category: 'bundle',
     });
     try {
-      if (previousBundleGraphResult) {
+      if (canIncrementallyBundle && previousBundleGraphResult) {
         internalBundleGraph = previousBundleGraphResult.bundleGraph;
         for (let changedAssetId of changedAssets.keys()) {
           // Copy over the whole node to also have correct symbol data
@@ -488,6 +487,7 @@ class BundlerRunner {
         this.api.storeResult(
           {
             bundleGraph: internalBundleGraph,
+            assetGraphBundlingVersion: graph.getBundlingVersion(),
             changedAssets: new Map(),
             assetRequests: [],
           },
@@ -512,7 +512,7 @@ class BundlerRunner {
     }
 
     let changedRuntimes = new Map();
-    if (!previousBundleGraphResult) {
+    if (!previousBundleGraphResult || !canIncrementallyBundle) {
       let namers = await this.config.getNamers();
       // inline bundles must still be named so the PackagerRunner
       // can match them to the correct packager/optimizer plugins.
@@ -564,6 +564,7 @@ class BundlerRunner {
     this.api.storeResult(
       {
         bundleGraph: internalBundleGraph,
+        assetGraphBundlingVersion: graph.getBundlingVersion(),
         changedAssets: new Map(),
         assetRequests: [],
       },
@@ -572,6 +573,7 @@ class BundlerRunner {
 
     return {
       bundleGraph: internalBundleGraph,
+      assetGraphBundlingVersion: graph.getBundlingVersion(),
       changedAssets: changedRuntimes,
       assetRequests,
     };