+++ packages/resolvers/glob/src/GlobResolver.js
@@ -222,7 +222,7 @@ export default (new Resolver({
       priority: 'sync',
     };
   },
-}): Resolver);
+}): Resolver<mixed>);
 
 function set(obj, path, value) {
   for (let i = 0; i < path.length - 1; i++) {