+++ packages/runtimes/hmr/src/HMRRuntime.js
@@ -61,4 +61,4 @@ export default (new Runtime({
       },
     };
   },
-}): Runtime);
+}): Runtime<mixed>);