+++ packages/runtimes/react-refresh/src/ReactRefreshRuntime.js
@@ -77,4 +77,4 @@ export default (new Runtime({
       }
     }
   },
-}): Runtime);
+}): Runtime<mixed>);