+++ packages/transformers/js/src/JSTransformer.js
@@ -386,6 +386,7 @@ export default (new Transformer({
 
     let env: EnvMap = {};
 
+    // $FlowFixMe
     if (!config?.inlineEnvironment) {
       if (options.env.NODE_ENV != null) {
         env.NODE_ENV = options.env.NODE_ENV;
@@ -397,6 +398,7 @@ export default (new Transformer({
     } else if (Array.isArray(config?.inlineEnvironment)) {
       for (let match of globMatch(
         Object.keys(options.env),
+        // $FlowFixMe
         config.inlineEnvironment,
       )) {
         env[match] = String(options.env[match]);
@@ -411,6 +413,7 @@ export default (new Transformer({
 
     let supportsModuleWorkers =
       asset.env.shouldScopeHoist && asset.env.supports('worker-module', true);
+    // $FlowFixMe
     let isJSX = Boolean(config?.isJSX);
     if (asset.isSource) {
       if (asset.type === 'ts') {
@@ -441,6 +444,7 @@ export default (new Transformer({
       module_id: asset.id,
       project_root: options.projectRoot,
       replace_env: !asset.env.isNode(),
+      // $FlowFixMe
       inline_fs: Boolean(config?.inlineFS) && !asset.env.isNode(),
       insert_node_globals:
         !asset.env.isNode() && asset.env.sourceType !== 'script',
@@ -450,9 +454,13 @@ export default (new Transformer({
       env,
       is_type_script: asset.type === 'ts' || asset.type === 'tsx',
       is_jsx: isJSX,
+      // $FlowFixMe
       jsx_pragma: config?.pragma,
+      // $FlowFixMe
       jsx_pragma_frag: config?.pragmaFrag,
+      // $FlowFixMe
       automatic_jsx_runtime: Boolean(config?.automaticJSXRuntime),
+      // $FlowFixMe
       jsx_import_source: config?.jsxImportSource,
       is_development: options.mode === 'development',
       react_refresh:
@@ -460,8 +468,11 @@ export default (new Transformer({
         !asset.env.isLibrary &&
         !asset.env.isWorker() &&
         !asset.env.isWorklet() &&
+        // $FlowFixMe
         Boolean(config?.reactRefresh),
+      // $FlowFixMe
       decorators: Boolean(config?.decorators),
+      // $FlowFixMe
       use_define_for_class_fields: Boolean(config?.useDefineForClassFields),
       targets,
       source_maps: !!asset.env.sourceMap,
@@ -474,10 +485,12 @@ export default (new Transformer({
       trace_bailouts: options.logLevel === 'verbose',
       is_swc_helpers: /@swc[/\\]helpers/.test(asset.filePath),
       standalone: asset.query.has('standalone'),
+      // $FlowFixMe
       inline_constants: config.inlineConstants,
       conditional_bundling: options.featureFlags.conditionalBundlingApi,
       hmr_improvements: options.featureFlags.hmrImprovements,
       computed_properties_fix: options.featureFlags.unusedComputedPropertyFix,
+      // $FlowFixMe
       magic_comments: Boolean(config?.magicComments),
       callMacro: asset.isSource
         ? async (err, src, exportName, args, loc) => {
@@ -1111,4 +1124,4 @@ export default (new Transformer({
 
     return [asset, ...macroAssets];
   },
-}): Transformer);
+}): Transformer<mixed>);