+++ packages/transformers/webextension/src/WebExtensionTransformer.js
@@ -433,4 +433,4 @@ export default (new Transformer({
     asset.meta.webextEntry = true;
     return [asset];
   },
-}): Transformer);
+}): Transformer<mixed>);