+++ packages/transformers/worklet/src/WorkletTransformer.js
@@ -18,4 +18,4 @@ export default (new Transformer({
     });
     return [asset];
   },
-}): Transformer);
+}): Transformer<mixed>);