+++ packages/configs/repl/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @atlaspack/config-repl
 
+## 2.13.24
+
+### Patch Changes
+
+- Updated dependencies [[`daaa768`](https://github.com/atlassian-labs/atlaspack/commit/daaa7688786772d7e3713b71c5bba6b89ec704aa)]:
+  - @atlaspack/transformer-react-refresh-wrap@2.14.20
+  - @atlaspack/transformer-inline-string@2.14.20
+  - @atlaspack/resolver-repl-runtimes@2.14.20
+  - @atlaspack/runtime-react-refresh@2.14.20
+  - @atlaspack/transformer-posthtml@2.14.20
+  - @atlaspack/transformer-postcss@2.14.20
+  - @atlaspack/transformer-babel@2.14.20
+  - @atlaspack/optimizer-terser@2.14.20
+  - @atlaspack/resolver-default@2.14.20
+  - @atlaspack/transformer-html@2.14.20
+  - @atlaspack/transformer-json@2.14.20
+  - @atlaspack/bundler-default@3.1.2
+  - @atlaspack/transformer-css@2.14.20
+  - @atlaspack/transformer-raw@2.14.20
+  - @atlaspack/transformer-js@3.3.10
+  - @atlaspack/namer-default@2.14.20
+  - @atlaspack/packager-html@2.15.2
+  - @atlaspack/packager-css@2.14.20
+  - @atlaspack/packager-raw@2.14.20
+  - @atlaspack/packager-js@2.18.2
+  - @atlaspack/runtime-browser-hmr@2.14.20
+  - @atlaspack/runtime-js@2.15.2
+  - @atlaspack/compressor-raw@2.13.22
+  - @atlaspack/reporter-dev-server-sw@2.14.20
+  - @atlaspack/reporter-json@2.14.20
+
+## 2.13.23
+
+### Patch Changes
+
+- Updated dependencies [[`13aef17`](https://github.com/atlassian-labs/atlaspack/commit/13aef177eea289a6e40d2113b5ec1ac9be18a33d)]:
+  - @atlaspack/runtime-js@2.15.1
+  - @atlaspack/bundler-default@3.1.1
+  - @atlaspack/packager-html@2.15.1
+  - @atlaspack/packager-js@2.18.1
+  - @atlaspack/runtime-react-refresh@2.14.19
+  - @atlaspack/transformer-js@3.3.9
+  - @atlaspack/transformer-react-refresh-wrap@2.14.19
+  - @atlaspack/transformer-html@2.14.19
+  - @atlaspack/optimizer-terser@2.14.19
+  - @atlaspack/packager-css@2.14.19
+  - @atlaspack/reporter-dev-server-sw@2.14.19
+  - @atlaspack/reporter-json@2.14.19
+  - @atlaspack/runtime-browser-hmr@2.14.19
+  - @atlaspack/transformer-babel@2.14.19
+  - @atlaspack/transformer-css@2.14.19
+  - @atlaspack/transformer-postcss@2.14.19
+  - @atlaspack/transformer-posthtml@2.14.19
+  - @atlaspack/resolver-default@2.14.19
+  - @atlaspack/compressor-raw@2.13.21
+  - @atlaspack/namer-default@2.14.19
+  - @atlaspack/packager-raw@2.14.19
+  - @atlaspack/resolver-repl-runtimes@2.14.19
+  - @atlaspack/transformer-inline-string@2.14.19
+  - @atlaspack/transformer-json@2.14.19
+  - @atlaspack/transformer-raw@2.14.19
+
 ## 2.13.22
 
 ### Patch Changes