+++ packages/core/feature-flags/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaspack/feature-flags
 
+## 2.19.2
+
+### Patch Changes
+
+- [#706](https://github.com/atlassian-labs/atlaspack/pull/706) [`1c7865a`](https://github.com/atlassian-labs/atlaspack/commit/1c7865a64451116d94015e248302435839d347c0) Thanks [@yamadapc](https://github.com/yamadapc)! - Clean-up feature-flag
+
+- [#707](https://github.com/atlassian-labs/atlaspack/pull/707) [`a0b959f`](https://github.com/atlassian-labs/atlaspack/commit/a0b959fbf61fc3f820ff03c7e8988945fe40a91a) Thanks [@yamadapc](https://github.com/yamadapc)! - Fix content key not found exceptions when bundling is aborted after a unsafe to incrementally bundle asset graph request
+
+## 2.19.1
+
+### Patch Changes
+
+- [#692](https://github.com/atlassian-labs/atlaspack/pull/692) [`13aef17`](https://github.com/atlassian-labs/atlaspack/commit/13aef177eea289a6e40d2113b5ec1ac9be18a33d) Thanks [@JakeLane](https://github.com/JakeLane)! - Add fallback behaviour when conditional bundle is missing
+
 ## 2.19.0
 
 ### Minor Changes