+++ packages/core/profiler/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaspack/profiler
 
+## 2.14.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/types-internal@2.15.2
+
+## 2.14.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/types-internal@2.15.1
+
 ## 2.14.15
 
 ### Patch Changes