+++ packages/utils/atlaspack-watcher-watchman-js/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaspack/watcher-watchman-js
 
+## 2.14.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/utils@2.17.2
+
+## 2.14.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/utils@2.17.1
+
 ## 2.14.18
 
 ### Patch Changes